// skybox/types.ts
export type FieldGradientLayerShaderBinding = {
  anchorCount: number;
  index: number;
  layer: Extract<SkyboxManifestLayer, { type: "field-gradient" }>;
  parameterPrefix: string;
};
// ...
export type FieldGradientUniformNodes = {
  amplitude: ReturnType<typeof uniform>;
  anchors: Array<{
    color: ReturnType<typeof uniform>;
    direction: ReturnType<typeof uniform>;
  }>;
  frequency: ReturnType<typeof uniform>;
  layerId: string;
  mode: ReturnType<typeof uniform>;
  power: ReturnType<typeof uniform>;
};
// ...
export type BuiltInWebGpuLayerAdapter<
  TType extends SkyboxManifestLayer["type"],
  TBinding,
  TUniforms,
> = WebGpuLayerAdapter<Extract<SkyboxManifestLayer, { type: TType }>, TBinding, TUniforms>;